# Improve lazy performance of `Data.Text.Lazy.inits`