Closes #562

The previous implementation, itself based on an earlier version of `Data.List.inits`, inherited the flaw that accessing the i-th element took quadratic time O(i²). This now takes linear time O(i) as expected.

The current version of `Data.List.inits` uses a banker's queue to obtain good performance when generating very long lists. For lazy text, consisting of a few big chunks, that benefit seems negligible. So I chose a simpler implementation.

Benchmarks included.

Quadratic growth before (the "2k" benchmarks take 4x the time of the "1k" benchmarks):

```
    Lazy.inits
      last 1k:      OK
        9.19 ms ± 475 μs
      last 2k:      OK
        46.7 ms ± 1.5 ms
      map-take1 1k: OK
        8.99 ms ± 798 μs
      map-take1 2k: OK
        47.0 ms ± 1.8 ms
```

Linear growth after (the "2k" benchmarks take twice the time of the "1k" benchmarks):

```
Lazy.inits
      last 1k:      OK
        46.8 μs ± 3.7 μs
      last 2k:      OK
        94.6 μs ± 1.8 μs
      map-take1 1k: OK
        66.7 μs ± 688 ns
      map-take1 2k: OK
        132  μs ± 6.2 μs
```