@Lysxia 2024-03-16
> Is our test coverage good enough to generate lazy Text with multiple chunks?

Yes, the `Arbitrary TL.Text` instance explicitly generates arbitrary chunks:

https://github.com/haskell/text/blob/240682e57dd93e793db48ea2aa98c4dfd7f89453/tests/Tests/QuickCheckUtils.hs#L98